+++ climada_petals/entity/exposures/crop_production.py
@@ -30,7 +30,6 @@ from matplotlib import pyplot as plt
 
 from climada import CONFIG
 from climada.entity import Exposures, INDICATOR_IMPF
-from climada.util.tag import Tag
 import climada.util.coordinates as u_coord
 
 logging.root.setLevel(logging.DEBUG)
@@ -363,9 +362,9 @@ class CropProduction(Exposures):
             value_tmp = np.squeeze(area_crop[irr_val]*hist_mean_dict[irr_val][idx_mean])
             value_tmp = np.nan_to_num(value_tmp) # replace NaN by 0.0
             exp.gdf['value'] += value_tmp
-        exp.tag = Tag(description="Crop production exposure from ISIMIP"
-                                  f" {CROP_NAME[crop]['print']} {irr}"
-                                  f" {yearrange[0]} {yearrange[-1]}")
+        exp.description=("Crop production exposure from ISIMIP"
+                        f" {CROP_NAME[crop]['print']} {irr}"
+                        f" {yearrange[0]} {yearrange[-1]}")
         exp.value_unit = 't/y' # input unit, will be reset below if required by user
         exp.crop = crop
         exp.ref_year = yearrange
@@ -484,8 +483,8 @@ class CropProduction(Exposures):
         exp.gdf['value'] = np.multiply(data_area.values, data_yield.values).flatten()
 
         exp.crop = crop_type
-        exp.tag = Tag(description=f"Annual crop production from {var_area} and {var_yield} for"
-                                  f" {exp.crop} from files {filename_area} and {filename_yield}")
+        exp.description=(f"Annual crop production from {var_area} and {var_yield} for"
+                         f" {exp.crop} from files {filename_area} and {filename_yield}")
         exp.value_unit = 't/y'
         try:
             rows, cols, ras_trans = u_coord.pts_to_raster_meta(
@@ -1032,6 +1031,9 @@ def normalize_with_fao_cp(exp_firr, exp_noirr, input_dir=None,
             exp_firr.set_value_to_usd(input_dir=input_dir)
         elif 'kcal' in unit or 'kcal' in exp_firr.value_unit:
             exp_firr.set_value_to_kcal(biomass=True)
+
+    exp_firr_norm.description = " ".join([exp_firr_norm.description or "", "normalized"])
+    exp_noirr_norm.description = " ".join([exp_noirr_norm.description or "", "normalized"])
 
     if return_data:
         return country_list, ratio, exp_firr_norm, exp_noirr_norm, \