+++ climada_petals/test/test_spamagrar_integr.py
@@ -37,12 +37,16 @@ class TestDefault(unittest.TestCase):
         self.assertIn('Lat. range: -55.375 to +71.125.', cm.output[0])
         self.assertIn('Lon. range: -179.125 to +179.958.', cm.output[1])
         self.assertIn("Total V_agg TA global: 1301919384722.2 USD.", cm.output[2])
-        self.assertTrue(ent.gdf.region_id.min() == 4)
-        self.assertTrue(ent.gdf.region_id[10000] == 246)
-        self.assertTrue(ent.gdf.region_id.max() == 894)
-        self.assertTrue(ent_select.value.sum() == 1878858118.0)
-        self.assertTrue(ent_select.region_id.min() == 208)
-        self.assertTrue(ent_select.region_id.max() == 208)
+        self.assertEqual(ent.gdf.region_id.min(), 4)
+        self.assertEqual(ent.gdf.region_id[10000], 246)
+        self.assertEqual(ent.gdf.region_id.max(), 894)
+        self.assertEqual(ent_select.value.sum(), 1878858118.0)
+        self.assertEqual(ent_select.region_id.min(), 208)
+        self.assertEqual(ent_select.region_id.max(), 208)
+        self.assertEqual(ent.spam_file, "spam2005V3r2_global_V_agg_TA.csv")
+        self.assertEqual(ent.description, 
+            "SPAM agrar exposure for variable V_agg and technology TA\n"
+            "all technologies together, ie complete crop")
 
     def test_suriname_pass(self):
         """Test country Suriname for default parameters:"""
@@ -53,8 +57,12 @@ class TestDefault(unittest.TestCase):
         self.assertIn('Lat. range: +1.875 to +5.958.', cm.output[0])
         self.assertIn('Lon. range: -58.042 to -54.042.', cm.output[1])
         self.assertIn("Total V_agg TA Suriname: 78879225.2 USD.", cm.output[2])
-        self.assertTrue(ent.gdf.region_id.min() == 740)
-        self.assertTrue(ent.gdf.region_id.max() == 740)
+        self.assertEqual(ent.gdf.region_id.min(), 740)
+        self.assertEqual(ent.gdf.region_id.max(), 740)
+        self.assertEqual(ent.spam_file, "spam2005V3r2_global_V_agg_TA.csv")
+        self.assertEqual(ent.description, 
+            "SPAM agrar exposure for variable V_agg and technology TA\n"
+            "all technologies together, ie complete crop")
 
     def test_zurich_pass(self):
         """Test admin 1 Zurich for default parameters:"""
@@ -66,8 +74,12 @@ class TestDefault(unittest.TestCase):
         self.assertIn('Lat. range: +47.208 to +47.625.', cm.output[0])
         self.assertIn('Lon. range: +8.375 to +8.875.', cm.output[1])
         self.assertIn("Total V_agg TA CHE Zurich: 56644555.1 USD.", cm.output[2])
-        self.assertTrue(ent.gdf.region_id.min() == 756)
-        self.assertTrue(ent.gdf.region_id.max() == 756)
+        self.assertEqual(ent.gdf.region_id.min(), 756)
+        self.assertEqual(ent.gdf.region_id.max(), 756)
+        self.assertEqual(ent.spam_file, "spam2005V3r2_global_V_agg_TA.csv")
+        self.assertEqual(ent.description, 
+            "SPAM agrar exposure for variable V_agg and technology TA\n"
+            "all technologies together, ie complete crop")
 
 class TestOtherVar(unittest.TestCase):
     """Test SPAM exposures based on other variables."""
@@ -83,8 +95,12 @@ class TestOtherVar(unittest.TestCase):
         self.assertIn('Lat. range: +45.875 to +47.792.', cm.output[0])
         self.assertIn('Lon. range: +6.042 to +10.375.', cm.output[1])
         self.assertIn("Total H TI CHE: 28427.1 Ha.", cm.output[2])
-        self.assertTrue(ent.gdf.region_id.min() == 756)
-        self.assertTrue(ent.gdf.region_id.max() == 756)
+        self.assertEqual(ent.gdf.region_id.min(), 756)
+        self.assertEqual(ent.gdf.region_id.max(), 756)
+        self.assertEqual(ent.spam_file, "spam2005V3r2_global_H_TI.csv")
+        self.assertEqual(ent.description, 
+            "SPAM agrar exposure for variable H and technology TI\n"
+            "irrigated portion of crop")
 
     def test_ucayali_pass(self):
         """Test admin 2 region Ucayali for non-default parameters:"""
@@ -98,6 +114,10 @@ class TestOtherVar(unittest.TestCase):
         self.assertIn('Lat. range: -8.625 to -6.042.', cm.output[0])
         self.assertIn('Lon. range: -76.125 to -74.208.', cm.output[1])
         self.assertIn("Total Y TA  Ucayali: 12298441.3 kg/Ha.", cm.output[2])
+        self.assertEqual(ent.spam_file, "spam2005V3r2_global_Y_TA.csv")
+        self.assertEqual(ent.description, 
+            "SPAM agrar exposure for variable Y and technology TA\n"
+            "all technologies together, ie complete crop")
 
 class TestInvalidInput(unittest.TestCase):
     """Test SPAM exposures based on invalid inputs."""